pyproject.toml
[tool.poetry]
name = "chaoying.dev"
version = "0.1.0"
description = ""
authors = ["Chaoying <chaunceywe@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
starlette = "^0.15.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
loguru = "^0.5.3"
tortoise-orm = {extras = ["asyncpg"], version = "^0.17.7"}
Jinja2 = "^3.0.1"
python-multipart = "^0.0.5"
markdown2 = "^2.4.1"
python-slugify = "^5.0.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
isort = "^5.9.3"
flake8 = "^3.9.2"
mypy = "^0.910"
ipython = "^7.27.0"
coverage = "^5.5"
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
